//created class OvernightPackage

+++ part2/TwoDayPackage.cpp
@@ -10,7 +10,7 @@
 #include "TwoDayPackage.h"
 using namespace std;
 
-TwoDayPackage::TwoDayPackage(const string &n, const string &add, int zipCode, double w, double cPO, int c)
+TwoDayPackage::TwoDayPackage(const string &n, const string &add, int zipCode, double w, double cPO, double c)
 	: Package(n, add, zipCode, w, cPO)
 {
 	extraCost = c;
